+++ src/main/java/com/lab/web/util/Validator.java
@@ -2,8 +2,8 @@ package com.lab.web.util;
 
 public class Validator {
 
-    public static int validateX(String value){
-        return Integer.parseInt(value);
+    public static double validateX(String value){
+        return Double.parseDouble(value);
     }
 
     public static double validateY(String value){
